moved initialization of AWS mobile client to utils
