+++ app/src/main/java/e/josephmolina/saywhat/MainScreen/MainActivity.java
@@ -21,6 +21,7 @@ import com.crashlytics.android.Crashlytics;
 
 import e.josephmolina.saywhat.Dialog.SayWhatDialog;
 import e.josephmolina.saywhat.R;
+import e.josephmolina.saywhat.Utils.Utils;
 import io.fabric.sdk.android.Fabric;
 
 
@@ -32,7 +33,7 @@ public class MainActivity extends AppCompatActivity implements SayWhatDialog.Dia
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         Fabric.with(this, new Crashlytics());
-        initializeAWSMobileClient();
+        Utils.initializeAWSMobileClient(this);
         controller = new MainController(this);
     }
 
@@ -87,31 +88,5 @@ public class MainActivity extends AppCompatActivity implements SayWhatDialog.Dia
     protected void onDestroy() {
         super.onDestroy();
         controller.cleanUp();
-    }
-
-    void initializeAWSMobileClient() {
-        AWSMobileClient.getInstance().initialize(this, new AWSStartupHandler() {
-            @Override
-            public void onComplete(AWSStartupResult awsStartupResult) {
-
-                //Make a network call to retrieve the identity ID
-                // using IdentityManager. onIdentityId happens upon success.
-                IdentityManager.getDefaultIdentityManager().getUserID(new IdentityHandler() {
-
-                    @Override
-                    public void onIdentityId(String s) {
-                        //The network call to fetch AWS credentials succeeded, the cached
-                        // user ID is available from IdentityManager throughout your app
-                        Log.d("MainActivity", "Identity ID is: " + s);
-                        Log.d("MainActivity", "Cached Identity ID: " + IdentityManager.getDefaultIdentityManager().getCachedUserID());
-                    }
-
-                    @Override
-                    public void handleError(Exception e) {
-                        Log.e("MainActivity", "Error in retrieving Identity ID: " + e.getMessage());
-                    }
-                });
-            }
-        }).execute();
     }
 }
